[SecurityBundle] Move format-dependent tests from SecurityExtensionTest
