--- src/Symfony/Bundle/SecurityBundle/Tests/DependencyInjection/Fixtures/php/invalid_check_path.php
@@ -1,16 +1,0 @@
-<?php
-
-$container->loadFromExtension('security', array(
-    'providers' => array(
-        'default' => array('id' => 'foo'),
-    ),
-
-    'firewalls' => array(
-        'some_firewall' => array(
-            'pattern' => '/secured_area/.*',
-            'form_login' => array(
-                'check_path' => '/some_area/login_check',
-            )
-        )
-    )
-));
