--- src/Symfony/Bundle/SecurityBundle/Tests/DependencyInjection/Fixtures/xml/invalid_check_path.xml
@@ -1,16 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<srv:container xmlns="http://symfony.com/schema/dic/security"
-               xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-               xmlns:srv="http://symfony.com/schema/dic/services"
-               xsi:schemaLocation="http://symfony.com/schema/dic/services http://symfony.com/schema/dic/services/services-1.0.xsd">
-
-    <config>
-        <provider name="default" id="foo" />
-
-        <firewall name="some_firewall" pattern="/secured_area/.*">
-            <form-login check-path="/some_area/login_check" />
-        </firewall>
-    </config>
-
-</srv:container>
