--- src/Symfony/Bundle/SecurityBundle/Tests/DependencyInjection/Fixtures/yml/invalid_check_path.yml
@@ -1,9 +1,0 @@
-security:
-    providers:
-        default: { id: foo }
-
-    firewalls:
-        some_firewall:
-            pattern: /secured_area/.*
-            form_login:
-                check_path: /some_area/login_check
